Sort all tags when sorting is requested.
Previously, we'd just sort the tags for the current file being examined.
However, this is not how it was meant to be.  The sorted format is
supposed to be binary searchable across all tags, not just for the
relevant file.  So we just wait to write all the content out until the
end.

Note: we still write the header out first so that we match ctags
behavior.