Embed the version string in rst2ctags.
Turns out pkg_resources is dog slow and would take on the order of 100ms
just to import it.  That's far too much time for a fast command line
tool.  So we flip the model and embed the string in the source and have
setup.py pull it out.

Thanks to @drmikehenry for figuring this out!